backend/requirements.txt
# ─────────────────────────────────────────────
# Web Framework & ASGI Server
# ─────────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.29.0

# ─────────────────────────────────────────────
# Database - Async SQLite with SQLAlchemy ORM
# ─────────────────────────────────────────────
sqlalchemy==2.0.29
aiosqlite==0.20.0

# ─────────────────────────────────────────────
# Data Validation
# ─────────────────────────────────────────────
pydantic==2.6.4
pydantic-settings==2.2.1

# ─────────────────────────────────────────────
# Authentication - JWT Tokens
# ─────────────────────────────────────────────
python-jose[cryptography]==3.3.0

# ─────────────────────────────────────────────
# Authentication - Password Hashing
# ─────────────────────────────────────────────
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# ─────────────────────────────────────────────
# File Uploads - Required for CSV import
# ─────────────────────────────────────────────
python-multipart==0.0.9

# ─────────────────────────────────────────────
# PDF Generation - Post-auction export
# ─────────────────────────────────────────────
reportlab==4.1.0

# ─────────────────────────────────────────────
# Environment Variables
# ─────────────────────────────────────────────
python-dotenv==1.0.1

# ─────────────────────────────────────────────
# Async File Handling
# ─────────────────────────────────────────────
aiofiles==23.2.1


